Use column_view instead of column for all APIs and fix serialization

See merge request nvspark/cudf!253

+++ java/src/test/java/ai/rapids/cudf/ReductionTest.java
@@ -47,7 +47,7 @@ class ReductionTest extends CudfTestBase {
     case INT32:
       return Scalar.fromInt((Integer) expectedObject);
     case INT64:
-      return Scalar.fromLong((Integer) expectedObject);
+      return Scalar.fromLong((Long) expectedObject);
     case FLOAT32:
       return Scalar.fromFloat((Float) expectedObject);
     case FLOAT64:
